requirements.txt
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djoser==2.1.0
fake-useragent==0.1.11
playwright==1.23.0
requests==2.28.1
psycopg2-binary==2.9.3
celery==5.2.6
redis==4.3.1
django-celery-beat==2.3.0
pyvirtualdisplay==3.0
validators==0.20.0
django-auditlog==2.1.0
